app/requirements.txt
# ==============================================================
# REQUIREMENTS.TXT
# Every line is a Python library this project depends on.
# Install all of them with: pip install -r requirements.txt
# ==============================================================

# --- Web framework & server ---
fastapi==0.115.0         # The API framework
uvicorn[standard]==0.30.6 # The ASGI server that runs FastAPI

# --- Database ---
sqlalchemy==2.0.35       # ORM — write Python, get SQL
asyncpg==0.29.0          # Async PostgreSQL driver
alembic==1.13.2          # Database migration tool

# --- Configuration ---
pydantic-settings==2.4.0 # Read .env into typed Python config class
python-dotenv==1.0.1     # Load .env file into environment

# --- Authentication & Security ---
passlib==1.7.4           # Password hashing utilities
bcrypt==4.2.0            # The bcrypt hashing algorithm
python-jose[cryptography]==3.3.0  # JWT token creation & verification

# --- Testing ---
pytest==8.3.3            # Test runner
httpx==0.27.2            # HTTP client used by pytest to test endpoints
pytest-asyncio==0.24.0   # Allows pytest to run async test functions
aiosqlite==0.20.0        # Async SQLite driver used for the test database